Literal port: DSP is BYTE-IDENTICAL to the original for all 8 words
Fix the root.48 bridge rounding: Lua computes (R18+1)*R13/(R14+1) and the
sum R11+R20 in floating point, then the int16 store truncates the WHOLE
sum toward zero. Computing the quotient in integer first dropped 1 on
negative ramps. With double arithmetic + a single toward-zero truncation
of the sum, the bridge matches exactly.

Result: 0 sample differences (maxDiff=0) vs the clean single-word
reference PCM for mama, labas, saule, žodis, ačiū, obuolys, laipsnių,
lietuva. The Java DSP is now a sample-accurate literal port of
voicesynth root.51.1/51/52/52.5/49/48/47/45/50/44.

+++ app/src/main/java/com/rosasoft/wintalker/engine/DiphoneSynth.java
@@ -418,8 +418,11 @@ public final class DiphoneSynth {
                 int delta = nv - last;             // R13 = R8[0]-R10[R9-1]
                 int span = n7 - n9;                // R14 = R7-R9
                 for (int k = 0; k < span; k++)     // for R18 = 0 .. R14-1 (inclusive)
-                    // R10[R9+R18] = R11 + (R18+1)*R13/(R14+1)
-                    emit[n9 + k] = (short) (last + (long) (k + 1) * delta / (span + 1));
+                    // R10[R9+R18] = R11 + (R18+1)*R13/(R14+1). Lua computes the term
+                    // and sum in floating point, then the int16 store truncates the
+                    // WHOLE sum toward zero (not the quotient first).
+                    emit[n9 + k] = (short) (long)
+                        (last + (double) (k + 1) * delta / (span + 1));
             }
         } else {
             emit = prev;                           // unvoiced: R5 stays = prev.buffer
